+++ src/test/java/webserver/http/HttpMethodTest.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
